Rewrote ntlmrelayx attack logic to be similar to protocol clients
